% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_marks}
\alias{extract_marks}
\title{Extract marks and metadata from CVR JSON}
\usage{
extract_marks(sessions, max_marks)
}
\arguments{
\item{sessions}{CVR JSON}

\item{max_marks}{Maximum number of vote choice that JSON could contain}
}
\value{
dataframe containing the extracted vote choices and meta data
}
\description{
Extract marks and metadata from CVR JSON
}
