% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readManifest.R
\name{read_manifest}
\alias{read_manifest}
\title{Simple wrapper around manifest json file}
\usage{
read_manifest(path = NULL, zipdir = NULL)
}
\arguments{
\item{path}{The path to data as a vector of strings, typically with a \code{".json"} file extension.
If the file is within an unzipped zip file, also specify \code{zipdir}.}

\item{zipdir}{if the json files are in a zipped file and you do not want to
unzip the whole thing, you can list a \code{zipdir} so that the path \code{{zipdir}/{path}}
corresponds to a file. Then the function will extract the file internally.}
}
\description{
Simple wrapper around manifest json file
}
\examples{
\dontrun{
 read_manifest("data-raw/json/BallotTypeContestManifest.json")
 read_manifest("data-raw/json/BallotTypeManifest.json")
 read_manifest("data-raw/json/CandidateManifest.json")
 read_manifest("data-raw/json/ContestManifest.json")
 read_manifest("data-raw/json/CountingGroupManifest.json")
 read_manifest("data-raw/json/DistrictManifest.json")
 read_manifest("data-raw/json/DistrictPrecinctPortionManifest.json")
 read_manifest("data-raw/json/DistrictTypeManifest.json")
 read_manifest("data-raw/json/ElectionEventManifest.json")
 read_manifest("data-raw/json/OutstackConditionManifest.json")
 read_manifest("data-raw/json/PartyManifest.json")
 read_manifest("data-raw/json/PrecinctManifest.json")
 read_manifest("data-raw/json/PrecinctPortionManifest.json")
 read_manifest("data-raw/json/TabulatorManifest.json")
 }
}
