% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCVR.R
\name{extract_cvr}
\alias{extract_cvr}
\title{Extract tabular data from Dominion CVR}
\usage{
extract_cvr(cvr = NULL, path = NULL, verbose = TRUE)
}
\arguments{
\item{cvr, path}{The data. Use \code{cvr} for a list output from \code{jsonlite::fromJSON},
or use \code{path} to give the \code{fromJSON} path directly.}
}
\description{
Extract tabular data from Dominion CVR
}
\examples{

 library(furrr)
 plan("multicore")
 extract_cvr(path = c("data-raw/json/CvrExport_42.json", "data-raw/json/CvrExport_24940.json"))

}
